fix: Rename jhipster emulation
