+++ .env.example
@@ -1,5 +1,5 @@
 NEXT_PUBLIC_BASE_URL="http://localhost:3000"
-NEXT_PUBLIC_API_BASE_URL="http://localhost:3000/api/jhipster-emulation"
+NEXT_PUBLIC_API_BASE_URL="http://localhost:3000/api/jhipster-mocks"
 
 # Use the following environment variables to show the environment name.
 NEXT_PUBLIC_DEV_ENV_NAME="staging"
