+++ src/pages/api/jhipster-mocks/account/change-password.ts
