+++ src/pages/api/jhipster-mocks/account/index.ts
