+++ src/pages/api/jhipster-mocks/account/reset-password/finish.ts
