+++ src/pages/api/jhipster-mocks/account/reset-password/init.ts
