+++ src/pages/api/jhipster-mocks/activate.ts
