+++ src/pages/api/jhipster-mocks/admin/users/[login].ts
