+++ src/pages/api/jhipster-mocks/authenticate.ts
