+++ src/pages/api/jhipster-mocks/register.ts
